requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.15
boto3==1.34.20
python-docx==1.1.0
pdfplumber==0.9.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
pytest==8.2.0
mangum==0.17.0

